requirements.txt
streamlit==1.31.0
pandas==2.2.0
numpy==1.26.0
scipy==1.12.0
openai==1.12.0
sqlalchemy==2.0.25
plotly==5.18.0
python-dateutil==2.8.2
python-dotenv==1.0.0
